Let cx be usable in server pages
