+++ src/lib/actions/action.ts
@@ -1,7 +1,8 @@
 import { FormEvent, useCallback, useMemo, useRef, useState } from 'react'
 import { usePromise } from './promise'
 import { Result } from './result'
-import { promiseFinally, useMounted } from './utility'
+import { promiseFinally } from './utility'
+import { useMounted } from './utility-hooks'
 
 type ActionResult<T> = [
   // InvokeAction => Invokes the action, call this in an event handler
